backend/requirements.txt
Django  >= 5.2
djangorestframework>=3.16
mysqlclient >= 2.1.0
PyJWT>=2.6.0
djangorestframework_simplejwt>=5.2.2
django-cors-headers>=4.3.1
django-filter>=25.1
requests>=2.0.0
beautifulsoup4 >= 4.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0